+++ .dockerignore
@@ -6,3 +6,4 @@ tmp
 .github
 LICENSE
 .vscode
+coverage
