+++ .gitignore
@@ -4,3 +4,4 @@ data/
 config.json
 tmp
 .vscode
+coverage
