+++ web/game/controller.go
@@ -1,0 +1,135 @@
+package game
+
+import (
+	"net/http"
+	"time"
+
+	"github.com/alarmfox/game-repository/web"
+)
+
+type Service interface {
+	Create(request *CreateRequest) (Game, error)
+	FindById(id int64) (Game, error)
+	Delete(id int64) error
+	Update(id int64, ug *UpdateRequest) (Game, error)
+	FindByInterval(i *web.IntervalParams, p *web.PaginationParams) ([]Game, int64, error)
+}
+type Controller struct {
+	service Service
+}
+
+func NewController(gs Service) *Controller {
+	return &Controller{service: gs}
+}
+
+func (gc *Controller) Create(w http.ResponseWriter, r *http.Request) error {
+
+	request, err := web.FromJsonBody[CreateRequest](r.Body)
+
+	if err != nil {
+		return err
+	}
+
+	g, err := gc.service.Create(&request)
+
+	if err != nil {
+		return web.MakeHttpError(err)
+	}
+
+	return web.WriteJson(w, http.StatusCreated, g)
+
+}
+
+func (gc *Controller) FindByID(w http.ResponseWriter, r *http.Request) error {
+
+	id, err := web.FromUrlParams[Key](r, "id")
+	if err != nil {
+		return err
+	}
+
+	g, err := gc.service.FindById(id.AsInt64())
+
+	if err != nil {
+		return web.MakeHttpError(err)
+	}
+
+	return web.WriteJson(w, http.StatusOK, g)
+
+}
+
+func (gc *Controller) Delete(w http.ResponseWriter, r *http.Request) error {
+
+	id, err := web.FromUrlParams[Key](r, "id")
+	if err != nil {
+		return err
+	}
+
+	if err := gc.service.Delete(id.AsInt64()); err != nil {
+		return web.MakeHttpError(err)
+	}
+	w.WriteHeader(http.StatusNoContent)
+	return nil
+}
+
+func (gc *Controller) Update(w http.ResponseWriter, r *http.Request) error {
+
+	id, err := web.FromUrlParams[Key](r, "id")
+	if err != nil {
+		return err
+	}
+
+	request, err := web.FromJsonBody[UpdateRequest](r.Body)
+	if err != nil {
+		return err
+	}
+
+	g, err := gc.service.Update(id.AsInt64(), &request)
+	if err != nil {
+		return web.MakeHttpError(err)
+	}
+
+	return web.WriteJson(w, http.StatusOK, g)
+}
+
+func (gc *Controller) List(w http.ResponseWriter, r *http.Request) error {
+	page, err := web.FromUrlQuery[Key](r, "page", 1)
+
+	if err != nil {
+		return err
+	}
+
+	pageSize, err := web.FromUrlQuery[Key](r, "pageSize", 10)
+
+	if err != nil {
+		return err
+	}
+
+	startDate, err := web.FromUrlQuery(r, "startDate", Interval(time.Now().Add(-24*time.Hour)))
+
+	if err != nil {
+		return err
+	}
+
+	endDate, err := web.FromUrlQuery(r, "endDate", Interval(time.Now()))
+
+	if err != nil {
+		return err
+	}
+
+	ip := web.IntervalParams{
+		Start: startDate.AsTime(),
+		End:   endDate.AsTime(),
+	}
+
+	pp := web.PaginationParams{
+		Page:     page.AsInt64(),
+		PageSize: pageSize.AsInt64(),
+	}
+
+	games, count, err := gc.service.FindByInterval(&ip, &pp)
+	if err != nil {
+		return web.MakeHttpError(err)
+	}
+
+	return web.WriteJson(w, http.StatusOK, web.MakePaginatedResponse(games, count, &pp))
+}
